+++ examples/shooter/src/states/gameOverState.h
@@ -3,7 +3,7 @@
 #include <SDL2/SDL.h>
 #include <stormengine2/gameStateMachine.h>
 
-#include "../gamepad.h"
+#include <stormengine2/input/gamepad.h>
 #include <stormengine2/states/gameState.h>
 
 #include <string>